pyproject.toml
[project]
name = "dataconfy"
version = "0.0.1"
license-files = ["LICENSE"]
authors = [
    {name = "Luca Bello", email = "lusgabello@gmail.com"}
]
description = "Persist dataclass-based app config/data to disk (YAML/JSON) following XDG conventions, with simple load/save API."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "platformdirs>=4.5.0",
    "pyyaml>=6.0.3",
]

[project.urls]
Homepage = "https://github.com/lucabello/dataconfy"
Repository = "https://github.com/lucabello/dataconfy"
Issues = "https://github.com/lucabello/dataconfy/issues"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pyright>=1.1.406",
    "ruff>=0.14.0",
]

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.ruff]
line-length = 99

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[lint.pydocstyle]
convention = "google"
